Attempted to build X-Mem for Windows on ARM with and without NEON. Without NEON, it compiles OK but gets link errors saying that a QSD8960 processor bug cannot be avoided. So this is currently a no-go. With NEON enabled, I get type issues in benchmark kernel functions. But this is of secondary importance if I cannot link for ARM at all. Apparently desktop apps for ARM Windows is also not officially supported...

+++ src/Timer.cpp
@@ -66,13 +66,13 @@ Timer::Timer() :
 	stop_tick = stop_timer();
 	_ticks_per_ms = static_cast<tick_t>((stop_tick - start_tick) / BENCHMARK_DURATION_MS);
 #endif
-	_ns_per_tick = 1/(static_cast<float>(_ticks_per_ms)) * 1e6;
+	_ns_per_tick = 1/(static_cast<float>(_ticks_per_ms)) * static_cast<float>(1e6);
 }
 
 tick_t Timer::get_ticks_per_ms() {
 	return _ticks_per_ms;
 }
 
-double Timer::get_ns_per_tick() {
+float Timer::get_ns_per_tick() {
 	return _ns_per_tick;
 }
